move geo type to sharedAPI constants, remove unused activity type from constants
