--- api/src/constants.ts
@@ -1,9 +1,0 @@
-export const VALID_GEOMETRY_TYPES = [
-  'GeometryCollection',
-  'MultiPolygon',
-  'MultiLineString',
-  'MultiPoint',
-  'Polygon',
-  'LineString',
-  'Point'
-];
